requirements.txt
discord.py==2.5.2
python-dotenv==1.0.1
pystray==0.19.5
Pillow==10.4.0
